requirements.txt
streamlit==1.32.0
numpy==1.26.0
matplotlib==3.8.0
Pillow==10.0.0
scikit-image==0.21.0
scipy==1.11.3